chore: restructured folder for clarity
