+++ next/src/app/about/page.tsx
@@ -3,7 +3,8 @@ export const dynamic = "force-dynamic";
 import qs from 'qs';
 import type { Metadata } from "next";
 import { notFound } from 'next/navigation';
-import { getStrapiData, blockRenderer, Block, getStrapiMetadata } from "@/app/utils/utils";
+import { getStrapiData, getStrapiMetadata } from "../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../utils/render"
 
 const path = "/api/about";
 const queryAbout = qs.stringify({
@@ -68,7 +69,7 @@ export default async function About() {
 			</div>
 		)
 	} catch(error) {
-		console.error("Strapi fetch error: page About");
+		console.error(`Page About. ${error}`);
 		throw error;
 	}
 }
