+++ next/src/app/page.tsx
@@ -3,7 +3,8 @@ export const dynamic = "force-dynamic";
 import qs from "qs";
 import { notFound } from 'next/navigation';
 import type { Metadata } from "next";
-import { getStrapiData, blockRenderer, Block, getStrapiMetadata } from "@/app/utils/utils"
+import { getStrapiData, getStrapiMetadata } from "../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../utils/render"
 
 const path = "/api/accueil";
 const queryHero = qs.stringify({
@@ -78,7 +79,7 @@ export default async function Home() {
 			</div>
 		)
 	} catch(error) {
-		console.error("Strapi fetch error: page Home");
+		console.error(`Page Home. ${error}`);
 		throw error;
 	}
 }
