+++ next/src/app/prototype/page.tsx
@@ -3,7 +3,8 @@ export const dynamic = "force-dynamic";
 import qs from 'qs';
 import type { Metadata } from "next";
 import { notFound } from 'next/navigation';
-import { getStrapiData, blockRenderer, Block, getStrapiMetadata } from "@/app/utils/utils";
+import { getStrapiData, getStrapiMetadata } from "../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../utils/render"
 
 const path = "/api/prototype-service";
 const queryPrototype = qs.stringify(
@@ -65,7 +66,7 @@ export default async function Prototype() {
 			</div>
 		)
 	} catch(error) {
-		console.error("Strapi fetch error: page Prototype");
+		console.error(`Page Prototype. ${error}`);
 		throw error;
 	}
 }
