+++ next/src/app/student/page.tsx
@@ -3,7 +3,8 @@ export const dynamic = "force-dynamic";
 import qs from 'qs';
 import { notFound } from 'next/navigation';
 import type { Metadata } from "next";
-import { getStrapiData, blockRenderer, Block, getStrapiMetadata } from "@/app/utils/utils";
+import { getStrapiData, getStrapiMetadata } from "../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../utils/render"
 
 const path = "/api/student";
 const queryStudent = qs.stringify({
@@ -43,7 +44,7 @@ export default async function Student() {
 			</div>
 		)
 	} catch(error) {
-		console.error("Strapi fetch error: page Student");
+		console.error(`Page Student. ${error}`);
 		throw error;
 	}
 }
