+++ next/src/ui/accordion.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react"
 import * as AccordionPrimitive from "@radix-ui/react-accordion"
-import { ChevronDownIcon } from "lucide-react"
+// import { ChevronDownIcon } from "lucide-react"
 
 import { cn } from "@/lib/utils"
